+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [v3.0.1](http://github.com/ndustrialio/contxt-sdk-js/tree/v3.0.1) (2021-01-11)
+
+**Changed**
+
+- Roles are tied to Project Environments instead of Projects
+- `Coordinator.roles#addProject` is now `Coordinator.roles#addProjectEnvironment`
+- `Coordinator.roles#removeProject` is now `Coordinator.roles#removeProjectEnvironment`
+- `Coordinator.users#addProject` is now `Coordinator.users#addProjectEnvironment`
+- `Coordinator.users#removeProject` is now `Coordinator.users#removeProjectEnvironment`
+
 ## [v3.0.0](http://github.com/ndustrialio/contxt-sdk-js/tree/v3.0.0) (2020-12-16)
 
 **Changed**